src/core/preprocessing_service.py
from spellchecker import SpellChecker

class PreprocessingService:
<<<<<<< HEAD
    """
    A service for cleaning and correcting text before analysis.
    """
    def __init__(self):
        self.spell = SpellChecker()
||||||| c46cdd8
    """
    A placeholder for the missing PreprocessingService.
    This class is intended to resolve an ImportError and allow the test suite to run.
    """
    def __init__(self, *args, **kwargs):
        pass
=======
    """A mock service for preprocessing text."""
    def __init__(self):
        """Initializes the mock PreprocessingService."""
        pass
>>>>>>> origin/main

<<<<<<< HEAD
    def correct_text(self, text: str) -> str:
        """
        Corrects spelling mistakes in the given text.
        """
        words = text.split()
        corrected_words = [self.spell.correction(word) if self.spell.correction(word) is not None else word for word in words]
        return " ".join(corrected_words)
||||||| c46cdd8
    def correct_text(self, text):
        """
        A placeholder method that returns the text as-is.
        """
        return text
=======
    @staticmethod
    def correct_text(text: str) -> str:
        """A mock text correction method."""
        # For the mock, just return the original text.
        return text
>>>>>>> origin/main
