requirements.txt
# Core Web & API
fastapi
uvicorn
gunicorn
slowapi
requests
httpx

# GUI
PyQt6
PyQt6-WebEngine
matplotlib

# Data & Parsing
pdfplumber
pypdfium2
python-docx
pandas
openpyxl
pyyaml
markdown
pytesseract

# AI & Machine Learning
torch
transformers
sentence-transformers
ctransformers
faiss-cpu
bitsandbytes
rank_bm25
accelerate
nltk
pyspellchecker

# Database
SQLAlchemy
rdflib
aiosqlite

# Security & Auth
passlib[bcrypt]
python-multipart
PyJWT
python-jose[cryptography]

# Numeric & Science
numpy
scipy

# Scheduling
apscheduler
tenacity

# Testing
pytest
pytest-mock
pytest-qt
pytest-asyncio